requirements.txt
psycopg2-binary==2.9.10
pandas
Faker
jupyter
ipykernel
python-dotenv
faker-commerce